requirements.txt
asgiref==3.4.1
Babel==2.9.1
backports.zoneinfo==0.2.1
Django==4.0
django-money==2.1
django-treebeard==4.5.1
djangorestframework==3.13.0
djangorestframework-recursive==0.1.2
py-moneyed==1.2
pytz==2021.3
sqlparse==0.4.2
